regx: move most functions from the fwd component to base
