+++ orte/mca/regx/base/Makefile.am
@@ -1,5 +1,7 @@
 #
 # Copyright (c) 2015-2018 Intel, Inc.  All rights reserved.
+# Copyright (c) 2018      Research Organization for Information Science
+#                         and Technology (RIST). All rights reserved.
 # $COPYRIGHT$
 #
 # Additional copyrights may follow
@@ -11,5 +13,6 @@ headers += \
         base/base.h
 
 libmca_regx_la_SOURCES += \
+        base/regx_base_default_fns.c \
         base/regx_base_frame.c \
         base/regx_base_select.c
